refactor: keep a single solution to manage preview
Previously, there were both bash and npm scripts that rely on several nodejs
script. For simplicity, we only have a single bash script that support all
existing options (with a single nodejs script).

The documentation has been improved
  - introduce `--help` in the bash script to display all available options
  - the README better explains how to generate the preview for both development
  and deployment

Bonus: activate Admonitions with github support in the README
