+++ README.adoc
@@ -1,4 +1,12 @@
 = New Bonita Documentation Site
+:icons: font
+ifdef::env-github[]
+:note-caption: :information_source:
+:tip-caption: :bulb:
+:important-caption: :heavy_exclamation_mark:
+:caution-caption: :fire:
+:warning-caption: :warning:
+endif::[]
 // External URIs:
 :url-antora: https://antora.org
 :url-asciidoctor: https://asciidoctor.org/
@@ -46,6 +54,13 @@ For more CLI options, see https://docs.antora.org/antora/2.3/cli/options/
 
 ==== Building for local developments
 
+[WARNING]
+====
+The following command https://docs.antora.org/antora/2.3/playbook/runtime-fetch/[fetch the documentation content repositories] each time they are run. +
+<<local-content>> or <<preview>> can prevent this to happen.
+====
+
+
 For a site that can be load from local disk
 ```bash
 npm run local:build-static
@@ -56,7 +71,8 @@ For a site to be served by the development <<http server>>
 npm run local:build
 ```
 
-===== Using local documentation content repository
+[[local-content]]
+===== Using local documentation content repositories
 
 This activates the https://docs.antora.org/antora/2.3/playbook/author-mode/[Antora author mode]. The local content of the
 repository is used instead of fetching the remote one.
@@ -79,11 +95,28 @@ root
 ```
 ====
 
+For more information about the `build-preview.bash` script, see later in this document
 ```bash
-npm run local:build-static:local-content
+./build-preview.bash --ci false  --type local --local-sources true
 ```
 
+[[preview]]
+==== Building for preview
 
+It is possible to use a script that let's generate the documentation partially or with specific settings for both local
+developments or Pull Request previews.
+
+This includes building the documentation using:
+
+- already checkout local documentation content repositories
+- single branch for all components
+- single branch of a single component
+- ....
+
+To know all available options, run
+```bash
+./build-preview.bash --help
+```
 
 ==== http server
